requirements.txt
pip==25.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
coverage==7.6.10
dnspython==2.7.0
Flask==3.1.0
Flask-PyMongo==3.0.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
PyJWT==2.10.1
pymongo==4.11
pytest==8.3.4
Werkzeug==3.1.3
python-dotenv
